rename: 스크립트들 namespace내부로 이동
