+++ Project1108/Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -54,6 +54,8 @@ MonoBehaviour:
   - DelayShowUserNames
   - ReprintUserName
   - PrintUserNames
+  - CollectOnPhoton
+  - RegenOnPhoton
   DisableAutoOpenWizard: 1
   ShowSettings: 0
   DevRegionSetOnce: 1
