--- Project1108/Assets/YJH/Scripts/JoystickYoo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9feb52de8ef53f44e926382fb3fe1e20
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
